chore: 统一 uint32_t 的 printf 格式串 (%lu / %lx)
newlib-nano 的 stdint 把 uint32_t typedef 成 'long unsigned int',
GCC 的 -Wformat= 因此对 "%u + uint32_t" 报 19 处 warning. 之前 M2-11
仓促修了 http_ota.c / oplog.c 那几处 — 加 (unsigned) cast 绕过, 但代码
库整体仍有 warning.

本次治根: 把所有 uint32_t 的格式 specifier 统一改 %u→%lu, %x→%lx, 同时
回退之前加的 (unsigned) cast. logger.c 内的 vsnprintf 已忽略 'l' 修饰
(line 97), va_arg 仍按 uint32_t 取, runtime 行为不变.

clean rebuild 0 warning.

+++ openload/proto/http_ota.c
@@ -212,7 +212,7 @@ static int parse_header_line(http_ota_priv_t *p)
     if (p->line_buf[0] == 0) {
         /* 空行: 头部结束 */
         p->state = HS_RECV_BODY;
-        OL_LOGI("http: body %u bytes (0=unknown)", (unsigned)p->body_total);
+        OL_LOGI("http: body %lu bytes (0=unknown)", p->body_total);
         return OL_OK;
     }
 
@@ -274,8 +274,8 @@ static int http_poll(ol_receiver_t *r)
 
     int wrc = ol_part_write(p->dst, p->write_off, buf, (uint32_t)n);
     if (wrc != OL_OK) {
-        OL_LOGE("http: flash write at off=%u: %s",
-                (unsigned)p->write_off, ol_strerror(wrc));
+        OL_LOGE("http: flash write at off=%lu: %s",
+                p->write_off, ol_strerror(wrc));
         return wrc;
     }
     p->write_off     += (uint32_t)n;
@@ -283,7 +283,7 @@ static int http_poll(ol_receiver_t *r)
 
     if (p->body_total && p->body_received >= p->body_total) {
         p->state = HS_DONE;
-        OL_LOGI("http: body received %u bytes", (unsigned)p->body_received);
+        OL_LOGI("http: body received %lu bytes", p->body_received);
         return 1;
     }
     return 0;
